Improved refactoring
